ethtool: avoid null pointer dereference
'$ scan-build make' reports:

Description: Array access (from variable 'arg') results in a null
pointer dereference
File: /git/ethtool/netlink/parser.c
Line: 782

Description: Dereference of null pointer (loaded from variable 'p')
File: /git/ethtool/netlink/parser.c
Line: 794

Both of these bugs are prevented by checking the input in
nl_parse_char_bitset(), which is called from nl_sset() via the kernel
callback, specifically for the parsing of the wake-on-lan options (-s wol).
None of the other functions in this file seem to have the issue of
deferencing data without checking for validity first. This could
"technically" allow nlctxt->argp to be NULL, and scan-build is limited in
it's ability to parse for bugs only at file scope in this case. This
particular bug should be unlikely to happen because the kernel
builds/parses the netlink structure before handing it to the application.
However in the interests of being able to run this scan-build tool
regularly, I'm still sending the initial version of this patch as I tried
several other ways to fix the bug with an earlier check for NULL in
nl_sset, but it won't prevent the scan-build error due to the file scope
problem.

CC: Michal Kubecek <mkubecek@suse.cz>
Fixes: 81a30f416ec7 ("netlink: add bitset command line parser handlers")
Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
---
v3: updated commit message with more nuance after feedback from ethtool
maintainer. I'd be open to fixing this a different way but this seemed
the most straight-forward with the smallest amount of code changed.
v2: original version