ethtool: merge uapi changes to implement _BITUL and friends
I was looking into some errors reported by the runtime sanitizers and
found a couple of places where (1 << 31) was being used. This is a shift
of a bit into the sign-bit of an integer. This is undefined behavior for
the C-specification, and can be easily fixed with using (1UL << 31)
instead. A better way to do this is to use the _BITUL() macro, which
already has the 1UL in it (see future patch in series).

Convert and sync with the same changes made upstream to the uapi file,
to implement ethtool uapi use of _BITUL() and friends.

NOTE:
This change will be followed by a larger conversion patch, but *this*
commit only has the UAPI file changes and the initial implementation to
keep the work separate from the application only changes.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
---
v3: switch to using uapi/linux/const.h defines _BITUL() and _BITULL(),
    and follow the uapi directions for merging uapi changes from the
    kernel. removed fsl_enet.c changes to next patch.
v2: original version