ethtool: fix leak of memory after realloc
cppcheck finds:
netlink/msgbuff.c:63:2: error: Memory leak: nbuff [memleak]
 return 0;
 ^

This is a misleading error, as the original buffer simply wasn't freed
correctly if realloc fails. Move the msgbuff_done() function up so we
can re-use the code upon realloc failure, and then simplify the rest of
this function to not have any branches, as we already know new_size is
larger than old_size, and since the realloc didn't fail, we know we have
a new buffer that needs to be unconditionally set up.

Suggested-by: Michal Kubecek <mkubecek@suse.cz>
Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
---
v3: refactor with comments taken into account and simplifying code to
handle error and clean up the common "it worked" path.
v2: first version posted