ethtool: fix bug and use standard string parsing
The parse_reset function was open-coding string parsing routines and can
be converted to using standard routines. It was trying to be tricky and
parse partial strings to make things like "mgmt" and "mgmt-shared"
strings mostly use the same code. This is better done with strncmp().

This also fixes an array overflow bug where it's possible for the for
loop to access past the end of the input "val" array, which is bad.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
---
v3: fix review comments and refactor code to be correct, tested even
more thoroughly this time with mix of good and bad arguments.
v2: original version